Add support for the cert and bearer token based authentication to the Kubernetes container driver.
Also update KubeVirt compute driver to utilize shared code instead of
having duplicated code live in two places.

Now shared code lives in libcloud.common.kubernetes module.

Also update existing test and add additional tests for verifying auth
params.
